% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\docType{data}
\name{chr_lens}
\alias{chr_lens}
\title{GRCm39 Chromosome Lengths}
\format{
An object of class \code{numeric} of length 22.
}
\usage{
chr_lens
}
\description{
GRCm39 Chromosome Lengths
}
\keyword{datasets}
