% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{clean_chr_factor}
\alias{clean_chr_factor}
\title{Standardized Positional Factor}
\usage{
clean_chr_factor(chr_vector)
}
\arguments{
\item{chr_vector}{A vector of chromosome labels.}
}
\description{
Ensures chromosomes are ordered correctly as factors for plotting.
}
