% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\docType{data}
\name{context_colors}
\alias{context_colors}
\title{Color Palette for Diets and Sexes}
\format{
An object of class \code{character} of length 6.
}
\usage{
context_colors
}
\description{
Color Palette for Diets and Sexes
}
\keyword{datasets}
