% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{ensure_dir}
\alias{ensure_dir}
\title{Ensure Directory Exists}
\usage{
ensure_dir(path)
}
\arguments{
\item{path}{The directory path to check/create.}
}
\description{
A helper to create nested directories if they don't already exist.
}
