% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hotspot.R
\name{find_differential_hotspots}
\alias{find_differential_hotspots}
\title{Find Differential Hotspots}
\usage{
find_differential_hotspots(density_data, raw_data, threshold = 10)
}
\arguments{
\item{density_data}{Combined density data for at least two groups.}

\item{raw_data}{Master trans-eQTL data frame containing phenotype/gene details.}

\item{threshold}{Minimum difference in count to consider a hotspot "differential".}
}
\value{
A data frame containing the traits mapping to the differential hotspot regions.
}
\description{
Identifies regions where the difference in eQTL density between two groups
exceeds a specified threshold.
}
