% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtl.R
\name{get_specificity}
\alias{get_specificity}
\title{Determine QTL Specificity}
\usage{
get_specificity(df_a, df_b, dist_kb = 4000)
}
\arguments{
\item{df_a}{A data frame containing QTL mapping results for the first group.}

\item{df_b}{A data frame containing QTL mapping results for the second group.}

\item{dist_kb}{Coordinate distance in kilobases to define shared peaks (default 4000).}
}
\value{
A list containing two data frames (a_final and b_final) with an added
\code{qtl_status} column.
}
\description{
This function identifies whether a QTL is "Shared" or "Specific" by comparing
two datasets (e.g., different diets or sexes). A QTL is considered shared if
the same phenotype has a peak on the same chromosome within a specified distance.
}
