% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\docType{data}
\name{GLOBAL_MAP}
\alias{GLOBAL_MAP}
\title{Global Coordinate Map for plotting}
\format{
An object of class \code{data.frame} with 22 rows and 5 columns.
}
\usage{
GLOBAL_MAP
}
\description{
Global Coordinate Map for plotting
}
\keyword{datasets}
