% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conserve.R
\name{plot.conserve_analysis}
\alias{plot.conserve_analysis}
\title{Plot SNP Conservation Analysis}
\usage{
\method{plot}{conserve_analysis}(x, trait_name = NULL, ...)
}
\arguments{
\item{x}{An object of class \code{conserve_analysis}.}

\item{trait_name}{Optional. If provided, plots the Manhattan for a specific trait instead of global hotspots.}

\item{...}{Extra arguments (ignored).}
}
\value{
A ggplot object.
}
\description{
Generates a Manhattan plot for the global hotspots.
}
