% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hotspot.R
\name{plot.hotspot_analysis}
\alias{plot.hotspot_analysis}
\title{Plot Hotspot Analysis}
\usage{
\method{plot}{hotspot_analysis}(x, ...)
}
\arguments{
\item{x}{An object of class \code{hotspot_analysis}.}

\item{...}{Extra arguments (passed to plotting functions).}
}
\value{
A list of ggplot objects.
}
\description{
Plot Hotspot Analysis
}
