% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hotspot.R
\name{plot_hotspots}
\alias{plot_hotspots}
\title{Plot Hotspot Density}
\usage{
plot_hotspots(density_data, title)
}
\arguments{
\item{density_data}{Data frame generated by \code{calculate_hotspot_density}.}

\item{title}{Plot title.}
}
\value{
A ggplot object.
}
\description{
Visualizes the trans-eQTL count across chromosomes using a line plot.
}
