% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conserve.R
\name{print.conserve_analysis}
\alias{print.conserve_analysis}
\title{Print SNP Conservation Analysis Result}
\usage{
\method{print}{conserve_analysis}(x, ...)
}
\arguments{
\item{x}{An object of class \code{conserve_analysis}.}

\item{...}{Extra arguments (ignored).}
}
\description{
Print SNP Conservation Analysis Result
}
