% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hotspot.R
\name{print.hotspot_analysis}
\alias{print.hotspot_analysis}
\title{Print Hotspot Analysis Result}
\usage{
\method{print}{hotspot_analysis}(x, ...)
}
\arguments{
\item{x}{An object of class \code{hotspot_analysis}.}

\item{...}{Extra arguments (ignored).}
}
\description{
Print Hotspot Analysis Result
}
