% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtl.R
\name{print.qtl_analysis}
\alias{print.qtl_analysis}
\title{Print QTL Analysis Result}
\usage{
\method{print}{qtl_analysis}(x, digits = 4, ...)
}
\arguments{
\item{x}{An object of class \code{qtl_analysis}.}

\item{digits}{Number of significant digits for numeric values (default 4).}

\item{...}{Extra arguments (passed to print).}
}
\description{
Print QTL Analysis Result
}
