% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conserve.R
\name{read_conserve_analysis}
\alias{read_conserve_analysis}
\title{Read SNP Conservation Analysis Results from CSV}
\usage{
read_conserve_analysis(base_out)
}
\arguments{
\item{base_out}{Path to the directory containing the saved SNP conservation CSVs.}
}
\value{
An object of class \code{conserve_analysis}.
}
\description{
Reconstructs a \code{conserve_analysis} object from saved CSV files.
}
