% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hotspot.R
\name{read_hotspot_analysis}
\alias{read_hotspot_analysis}
\title{Read Hotspot Analysis Results from CSV}
\usage{
read_hotspot_analysis(output_path)
}
\arguments{
\item{output_path}{Path to the directory containing the saved hotspot analysis CSVs.}
}
\value{
An object of class \code{hotspot_analysis}.
}
\description{
Reconstructs a \code{hotspot_analysis} object from saved CSV files.
}
