% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{research_dir}
\alias{research_dir}
\title{Get Research Data Directory}
\usage{
research_dir()
}
\value{
The absolute path to the main data directory.
}
\description{
Resolves the path to the main research directory based on the operating system.
Mac: "/Volumes", Windows: "W:".
}
