% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{run_analysis}
\alias{run_analysis}
\title{Run Analysis Workflow}
\usage{
run_analysis(basename, ...)
}
\arguments{
\item{basename}{The basename of the workflow (e.g., "qtl", "hotspot", "conserve").}

\item{...}{Additional arguments passed to \code{source()}.}
}
\description{
Executes an analysis script from the package's \code{inst/scripts} directory.
}
