% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conserve.R
\name{run_trait_conservation}
\alias{run_trait_conservation}
\title{Run SNP Conservation Analysis for a Trait}
\usage{
run_trait_conservation(input_path, trait_name)
}
\arguments{
\item{input_path}{Path to the input SNP CSV file.}

\item{trait_name}{Name of the trait being analyzed.}
}
\value{
A data frame of prioritized SNPs.
}
\description{
Prioritizes SNPs for a specific trait by querying phastCons conservation scores
and incorporating SIFT scores if available.
}
