% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conserve.R
\name{summary.conserve_analysis}
\alias{summary.conserve_analysis}
\title{Summary of SNP Conservation Analysis}
\usage{
\method{summary}{conserve_analysis}(object, ...)
}
\arguments{
\item{object}{An object of class \code{conserve_analysis}.}

\item{...}{Extra arguments (ignored).}
}
\value{
A list of trait-level and global summaries.
}
\description{
Summary of SNP Conservation Analysis
}
