% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hotspot.R
\name{summary.hotspot_analysis}
\alias{summary.hotspot_analysis}
\title{Summary of Hotspot Analysis}
\usage{
\method{summary}{hotspot_analysis}(object, ...)
}
\arguments{
\item{object}{An object of class \code{hotspot_analysis}.}

\item{...}{Extra arguments (ignored).}
}
\value{
A list of differential hotspot traits.
}
\description{
Summary of Hotspot Analysis
}
