% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtl.R
\name{summary.qtl_analysis}
\alias{summary.qtl_analysis}
\title{Summary of QTL Analysis}
\usage{
\method{summary}{qtl_analysis}(object, ...)
}
\arguments{
\item{object}{An object of class \code{qtl_analysis}.}

\item{...}{Extra arguments (ignored).}
}
\value{
A data frame containing the QTL specificity summary.
}
\description{
Summary of QTL Analysis
}
