% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\docType{package}
\name{sysgenAnalysis}
\alias{sysgenAnalysis-package}
\alias{sysgenAnalysis}
\title{sysgenAnalysis: Systems Genetics Analysis for Attie Lab}
\description{
Routines for systems analysis, including trans-eQTL hotspot density calculation and visualization.
}
\author{
\strong{Maintainer}: Brian Yandell \email{brian.yandell@wisc.edu}

}
\keyword{internal}
