% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\docType{data}
\name{TOTAL_LEN}
\alias{TOTAL_LEN}
\title{Total Length of the genome in Mb}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
TOTAL_LEN
}
\description{
Total Length of the genome in Mb
}
\keyword{datasets}
