% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\docType{data}
\name{trait_class_colors}
\alias{trait_class_colors}
\title{Color Palette for Trait Classes}
\format{
An object of class \code{character} of length 4.
}
\usage{
trait_class_colors
}
\description{
Color Palette for Trait Classes
}
\keyword{datasets}
