% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtl.R
\name{plot.qtl_analysis}
\alias{plot.qtl_analysis}
\title{Plot QTL Analysis}
\usage{
\method{plot}{qtl_analysis}(x, ...)
}
\arguments{
\item{x}{An object of class \code{qtl_analysis}.}

\item{...}{Extra arguments (passed to ggsave if saving).}
}
\value{
A ggplot object.
}
\description{
Plot QTL Analysis
}
