% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hotspot.R
\name{plot_differential_hotspots}
\alias{plot_differential_hotspots}
\title{Plot Differential Hotspots}
\usage{
plot_differential_hotspots(density_data, group_names, title)
}
\arguments{
\item{density_data}{Combined density data for two groups.}

\item{group_names}{A vector of two group names (e.g., c("Female", "Male")).}

\item{title}{Plot title.}
}
\value{
A ggplot object.
}
\description{
Creates a bar plot showing the difference in eQTL density between two groups.
}
